requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
google-api-python-client==2.127.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
msal==1.28.0
httpx==0.27.0
pydantic==2.9.2
pydantic-settings==2.4.0
